requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
python-dotenv>=1.0.0
